validator functionality added
