+++ src/main/java/org/ehrbase/fhirbridge/camel/route/BundleRoutes.java
@@ -4,9 +4,9 @@ import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.RouteBuilder;
 import org.ehrbase.fhirbridge.camel.FhirBridgeConstants;
-import org.ehrbase.fhirbridge.fhir.bundle.BloodGasPanelBundleValidator;
+import org.ehrbase.fhirbridge.fhir.bundle.BundleValidator.BloodGasPanelBundleValidator;
 import org.ehrbase.fhirbridge.fhir.bundle.BloodGasPanelConverter;
-import org.ehrbase.fhirbridge.fhir.bundle.DiagnosticReportLabBundleValidator;
+import org.ehrbase.fhirbridge.fhir.bundle.BundleValidator.DiagnosticReportLabBundleValidator;
 import org.ehrbase.fhirbridge.fhir.bundle.DiagnosticReportLabConverter;
 import org.ehrbase.fhirbridge.fhir.common.Profile;
 import org.ehrbase.fhirbridge.fhir.util.FhirUtils;
